requirements.txt
google-adk
google-cloud-aiplatform
a2a-sdk
python-dateutil